Blockfrost.Client.Core

Description

Core shared by clients

Documentation

data BlockfrostError

Constructors

BlockfrostError Text

BlockfrostBadRequest Text

BlockfrostTokenMissing Text

BlockfrostNotFound

BlockfrostIPBanned

BlockfrostUsageLimitReached

BlockfrostFatal Text

ServantClientError ClientError

data Paged

Pagination parameters

Constructors

Paged

Fields

countPerPage :: Int

Count of results per page

pageNumber :: Int

Page number

data SortOrder

Constructors

Ascending

Descending

asc :: SortOrder

Ascending shortcut

def :: Default a => a

The default value for this type.

desc :: SortOrder

Descending shortcut

fromServantClientError :: ClientError -> BlockfrostError

newEnvByProject :: Project -> IO ClientEnv

page :: Int -> Paged

Default Paged at specific page number

paged :: Int -> Int -> Paged

Construct Paged from page size and page number.

Throws error on invalid values.

allPages :: Monad m => (Paged -> m [a]) -> m [a]

Query all results, until we get less than maximum items per page.

nextPage :: Paged -> Paged

Increment page number

projectFromEnv :: IO Project

Read file according to BLOCKFROST_TOKEN_PATH environment variable name.

projectFromEnv' :: String -> IO Project

Read file according to environment variable name.

projectFromFile :: FilePath -> IO Project

Read file with token and turn it into Project Expects tokens prefixed with environment, e.g. testnetA3C2E...

Orphan instances

ToMultipart Tmp Form

Instance details

Methods

toMultipart :: Form -> MultipartData Tmp